Enhance modals and add mobile-friendly product accordion component
- Added scroll behavior to various modals for better usability.
- Updated modal body styles to allow for overflow scrolling.
- Introduced ProductsAccordionListMobile component for improved mobile display of products.
- Updated ProductsPage to utilize the new accordion component.
- Added French translation for phone header in supplier details.
