+++ src/components/Intranet/Suppliers/SupplierEditModal.tsx
@@ -186,6 +186,7 @@ export default function SupplierEditModal({
             onOpenChange={(open) => {
                 if (!open) handleClose();
             }}
+            scrollBehavior="inside"
             className="sm:max-w-2xl w-full max-w-full m-0 sm:m-auto"
             style={{ height: "100%", maxHeight: "100dvh" }}
         >
