+++ src/components/ui/Form/AddFormModal.tsx
@@ -296,6 +296,7 @@ export default function AddFormModal({
                 isDismissable={false}
                 isKeyboardDismissDisabled={true}
                 size="2xl"
+                scrollBehavior="inside"
             >
                 <ModalContent>
                     {(onClose) => (
@@ -303,7 +304,7 @@ export default function AddFormModal({
                             <ModalHeader className="flex flex-col gap-1">
                                 {title}
                             </ModalHeader>
-                            <ModalBody>
+                            <ModalBody className="max-h-[60vh] overflow-y-auto">
                                 <Form
                                     id="add-form"
                                     onSubmit={handleSubmit}
